Popup windows by SweetAlert2

+++ src/components/Pin.js
@@ -1,7 +1,8 @@
 import React from 'react';
-
+import Swal from 'sweetalert2';
+import withReactContent from 'sweetalert2-react-content';
 import '../styles/pin_styles.css';
-// import OpenPin from './OpenPin.js';
+const MySwal = withReactContent(Swal);
 
 function checkSize(event) {
   const image = event.target;
@@ -12,13 +13,24 @@ function checkSize(event) {
   }
   image.style.opacity = 1;
 }
-//todo: DELETION
 function deletePin(pinDetails, deletePin) {
-  // console.log(pinDetails);
-  const pin_data = {
-    ...pinDetails,
-  };
-  deletePin(pin_data);
+  MySwal.fire({
+    title: 'Are you sure?',
+    text: 'Confirm deletion of the Pin',
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#e60023',
+    confirmButtonText: 'Yes, delete it!',
+  }).then((result) => {
+    if (result.isConfirmed) {
+      const pin_data = {
+        ...pinDetails,
+      };
+      deletePin(pin_data);
+      MySwal.fire('Deleted!', 'Pin has been deleted.', 'success');
+    }
+  });
 }
 
 function openPin(pinDetails, openPin) {
@@ -36,7 +48,6 @@ function Pin(props) {
       <div className='pin_modal'>
         <div className='modal_head'>
           <div className='save_card'>♡</div>
-          {/* <div className='delete_card'>⌫</div> */}
         </div>
 
         <div className='modal_foot'>
